backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
aiohttp>=3.9.0
scikit-learn>=1.4.0
numpy>=1.26.0

# Pydantic email validation
pydantic[email]>=2.0.0

# Auth dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0